Update Bash script to exit on empty argument.

+++ delete_filesystem.bash
@@ -2,6 +2,13 @@
 
 name=$1
 
+# Exit on empty name
+if [[ -z "${name// }" ]]; then
+	exit 1
+fi
+
+set +e # ignore error from umount, if FS is already unmounted
 umount -f /$name
+set -e
 rm -rf ~/$name.ext4
 rm -rf /$name
